requirements.txt
python-dotenv
python-telegram-bot
requests
garminconnect
garth
langchain-google-genai
pydantic